feat(m27 BL-112): leaf-extract unified_pipeline_enhanced.py (3806 -> 3192)
Clears the last test_file_size_limits red WITHOUT touching the hot
path. EnhancedPipeline (2865L, 59 methods incl. step_absorb,
step_quality, _run_pipeline_locked, DAG runner) and main stay
BYTE-IDENTICAL. Only standalone helpers / constants / classes move
to flat _pipeline_* SIBLINGS (same dir, so Path(__file__)-derived
path constants are unchanged) and are re-exported, so the ovp /
ovp-pipeline entrypoints and all 19 external importers (8 imported
symbols) are untouched.

Moved symbols (verbatim, no logic change):
- _pipeline_constants.py: all module constants (BASE_PIPELINE_STEPS,
  OPTIONAL_PIPELINE_STEPS, STEP_ALIASES, STAGE_CACHE policies,
  STAGE_ALGORITHM_VERSIONS, ENV paths, SCRIPTS_DIR, VAULT_DIR,
  PROJECT_ROOT/SRC, PIPELINE_STEP_CHOICES, ...)
- _pipeline_env.py: init_env_file, check_environment, _load_env,
  _check_api_key, detect_pinboard_processor,
  parse_pinboard_frontmatter
- _pipeline_plan.py: build_execution_plan, pipeline_steps,
  normalize_step_name
- _pipeline_runtime.py: TransactionManager, PipelineLogger
- _pipeline_utils.py: _get_version, _extract_json_suffix, _safe_int

Acyclic: constants <- env/plan; runtime/utils independent; main
imports all; no leaf imports main. Leaves take only the precise
absolute imports their moved code uses (NOT the original
script-fallback try/except shim - replicating that would add
silent-import sites). main re-exports EXPLICITLY (no star) and
prunes the 6 imports whose only consumers moved out
(looks_like_vault_dir plus 5 txn helpers), staying ruff
net-neutral (1 pre-existing StepContractError F401, unchanged).
_get_version's importlib version probe is a verbatim relocation of
the already-allowlisted unified_pipeline_enhanced.py:244
silent-import; LEGACY entry relocated to _pipeline_utils.py:44
(net-zero ratchet, same debt).

Deep EnhancedPipeline god-class decomposition is filed separately
as BL-112b (Later/P2) - explicitly NOT done here per scope.

125-pass hot-path/fitness/silent-import gate green
(test_runtime_paths, test_pipeline_data_contracts,
test_e2e_acceptance, test_architecture_fitness,
test_no_silent_imports). ruff: leaves clean, main net-neutral;
mypy net-neutral (3 leaf errors verbatim-carried from original).
